ESI_Model_Modular/Emer_API/requirements.txt
fastapi==0.109.0
uvicorn==0.25.0
pydantic==2.5.3
python-dotenv==1.0.0
torch==2.1.2
transformers==4.36.2
scikit-learn==1.3.2
langchain-groq==0.1.0
langchain-core==0.1.12
pandas==2.1.4
numpy==1.26.3